+++ plugins/module_utils/ibm_svc_utils.py
@@ -20,7 +20,7 @@ from ansible.module_utils.urls import open_url
 from ansible.module_utils.six.moves.urllib.parse import quote
 from ansible.module_utils.six.moves.urllib.error import HTTPError
 
-COLLECTION_VERSION = "3.0.0"
+COLLECTION_VERSION = "3.1.0"
 TIMEOUT = 600
 
 
@@ -108,7 +108,8 @@ def is_feature_supported(feature, version):
     # Feature or CLI to version mapping.
     # Example: First mapping shows that chvolume was introduced in SVC version 9.1.0.0
     feature_version_mapping = {
-        'chvolume': '9.1.0.0'
+        'chvolume': '9.1.0.0',
+        'certstore_cmds': '9.1.0.0'
     }
 
     return is_supported_version(version, feature_version_mapping[feature])
@@ -232,11 +233,11 @@ class IBMSVCRestApi(object):
             self.log('_svc_rest: httperror %s', str(e))
             r['code'] = e.getcode()
             r['out'] = e.read()
-            r['err'] = "HTTPError %s", str(e)
+            r['err'] = f"HTTPError {str(e)}"
             return r
         except Exception as e:
             self.log('_svc_rest: exception : %s', str(e))
-            r['err'] = "Exception %s", str(e)
+            r['err'] = f"Exception {str(e)}"
             return r
 
         try:
@@ -345,8 +346,12 @@ class IBMSVCRestApi(object):
         self.log("svc_run_command rest=%s", rest)
 
         if rest['err']:
-            msg = rest
-            self.module.fail_json(msg=msg)
+            self.log("[ERROR]: Unable to connect via REST: %s", rest['err'])
+            unreachable = False
+            if "Operation timed out" in rest['err']:
+                unreachable = True
+                self.module.exit_json(msg=rest, unreachable=unreachable)
+            self.module.fail_json(msg=rest)
             # Aborts
 
         # Might be None
@@ -391,6 +396,11 @@ class IBMSVCRestApi(object):
 
         # Fail for anything else
         if rest['err']:
+            self.log("[ERROR]: Unable to connect via REST: %s", rest['err'])
+            unreachable = False
+            if "Operation timed out" in rest['err']:
+                unreachable = True
+                self.module.exit_json(msg=rest, unreachable=unreachable)
             self.module.fail_json(msg=rest)
             # Aborts